+++ relay-server/src/utils/managed_envelope.rs
@@ -326,7 +326,7 @@ impl ManagedEnvelope {
         if self.context.summary.profile_quantity > 0 {
             self.track_outcome(
                 outcome,
-                if self.context.summary.profile_counted_as_processed {
+                if self.context.summary.event_metrics_extracted {
                     DataCategory::ProfileIndexed
                 } else {
                     DataCategory::Profile