+++ relay-server/src/utils/rate_limits.rs
@@ -143,11 +143,6 @@ pub struct EnvelopeSummary {
     /// Whether the envelope contains an event which already had the metrics extracted.
     pub event_metrics_extracted: bool,
 
-    /// Whether profiles in the envelope have been counted towards `DataCategory::Profile`.
-    ///
-    /// If `true`, count them towards `DataCategory::ProfileIndexed` instead.
-    pub profile_counted_as_processed: bool,
-
     /// The payload size of this envelope.
     pub payload_size: usize,
 }
@@ -172,10 +167,6 @@ impl EnvelopeSummary {
 
             if *item.ty() == ItemType::Transaction && item.metrics_extracted() {
                 summary.event_metrics_extracted = true;
-            }
-
-            if *item.ty() == ItemType::Profile && item.profile_counted_as_processed() {
-                summary.profile_counted_as_processed = true;
             }
 
             // If the item has been rate limited before, the quota has been consumed and outcomes
@@ -519,7 +510,7 @@ where
 
             // It makes no sense to store profiles without transactions, so if the event
             // is rate limited, rate limit profiles as well.
-            let profile_category = if summary.profile_counted_as_processed {
+            let profile_category = if summary.event_metrics_extracted {
                 DataCategory::ProfileIndexed
             } else {
                 DataCategory::Profile
@@ -562,7 +553,7 @@ where
             let item_scoping = scoping.item(DataCategory::Profile);
             let profile_limits = (self.check)(item_scoping, summary.profile_quantity)?;
             enforcement.profiles = CategoryLimit::new(
-                if summary.profile_counted_as_processed {
+                if summary.event_metrics_extracted {
                     DataCategory::ProfileIndexed
                 } else {
                     DataCategory::Profile